Define instance method possible? for PhoneNumber.
This method will check if the parsed phone number instance is a possible phone number.
If an input country code was passed at the time of parsing the number or if a default
country code existed, this will be used when determining if the parsed phone number is
valid. If none of those are passed or set, the default ZZ will be used most likely
resulting in the parsed number not being possible.