Add full_national and full_international instance methods to PhoneNumber.
Unfortunately, there doesn't seem to be a straightforward way to modify the
preferred extension prefix, so my best approach to doing a full format method that
includes the extension is to append it to the formatted number without it.
It seems that the default extension prefix is ext., but that may not be
what we want. For now, I'm defaulting to ; which is the one used by Phonelib,
but perhaps it may be best to customize it.